+++ .eslintrc.js
@@ -7,6 +7,7 @@ module.exports = {
     "eslint:recommended",
     "plugin:flowtype/recommended",
     "plugin:react/recommended",
+    "plugin:prettier/recommended",
     "prettier",
     "prettier/flowtype",
     "prettier/react",
@@ -32,6 +33,9 @@ module.exports = {
     __PATH_PREFIX__: true,
     __BASE_PATH__: true,
     __ASSET_PREFIX__: true,
+    dd: "readonly",
+    dump: "readonly",
+    clipboardy: "readonly",
   },
   rules: {
     "arrow-body-style": [
@@ -39,14 +43,9 @@ module.exports = {
       "as-needed",
       { requireReturnForObjectLiteral: true },
     ],
-    "no-unused-expressions": [
-      "error",
-      {
-        allowTaggedTemplates: true,
-      },
-    ],
-    "consistent-return": ["error"],
-    "filenames/match-regex": ["error", "^[a-z-\\d\\.]+$", true],
+    "no-unused-expressions": "off",
+    "@typescript-eslint/no-unused-expressions": "error",
+    // "filenames/match-regex": ["error", "^[a-z-\\d\\.]+$", true],
     "no-console": "off",
     "no-inner-declarations": "off",
     "prettier/prettier": "error",
@@ -103,10 +102,10 @@ module.exports = {
         // This ensures all interfaces are named with an I as a prefix
         // e.g.,
         // interface IFoo {}
-        "@typescript-eslint/interface-name-prefix": [
-          "error",
-          { prefixWithI: "always" },
-        ],
+        // "@typescript-eslint/interface-name-prefix": [
+        //   "error",
+        //   { prefixWithI: "always" },
+        // ],
         "@typescript-eslint/no-empty-function": "off",
         // This ensures that we always type the return type of functions
         // a high level focus of our TS setup is typing fn inputs and outputs.