+++ src/constants.ts
@@ -1,3 +1,0 @@
-export const CREATED_NODE_IDS: string = `WPGQL-created-node-ids`
-export const LAST_COMPLETED_SOURCE_TIME: string = `WPGQL-last-completed-source-time`
-export const MD5_CACHE_KEY: string = `introspection-node-query-md5`